requirements.txt
Django~=5.1.3
pip~=24.3.1
pillow~=11.0.0